Move the CFLAGS to before config.h
./waf configure cannot handle arbitrary CFLAGS going before the
config_chk() calls. An example disruption is when waf checks for
SHA256 capabilities. It creates a sample C file with an unused
variable p=...(SHA256); The compiler, with -Wall, observes p as
an unused-variable and throws an error- causing waf to report
that the system does not have SHA256 capabilities. Since we're
just checking capbilities, debug and optimization options should
only be included in our program's compilation options and not the
configuration process.